% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adVeloc.r
\name{adVeloc}
\alias{adVeloc}
\title{Advection velocities in porous medium}
\usage{
adVeloc(uRef, pRef, p, solid)
}
\arguments{
\item{uRef}{Reference advection velocity (e.g. at a certain depth).}

\item{pRef}{Porosity corresponding to \code{uRef}.}

\item{p}{Vector of porosities for which advection velocities are to be returned.}

\item{solid}{Logical. Should velocities be returned for solids or liquid?}
}
\value{
Vector of the same length as \code{p}.
}
\description{
Yields advection velocities of a solid or liquid component in a porous
medium to fulfil the constraint of mass conservation (continuity eqn.).
}
\note{
The underlying principle of continuity is also known as steady-state
  compaction; see \code{ReacTran::setup.compaction.1D}.
}
\examples{
z= 0:10
p= seq(0.99, 0.95, length.out=length(z))
plot(z, adVeloc(uRef=1, pRef=p[length(p)], p=p, solid=TRUE), type="l")
lines(z, adVeloc(uRef=1, pRef=p[length(p)], p=p, solid=FALSE), lty=2)
legend("top", bty="n", lty=1:2, legend=c("solid","liquid"))
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

