% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorQual.r
\name{colorQual}
\alias{colorQual}
\title{Assign colors to nominal values}
\usage{
colorQual(x, classes, colors, warn = TRUE, NAcolor = "transparent")
}
\arguments{
\item{x}{Vector of character strings.}

\item{classes}{Vector of character strings defining the class names.
Typically, this would be \code{unique(x)}.}

\item{colors}{Vector of colors corresponding to \code{classes}.}

\item{warn}{Warn if \code{x} contains elements not present in \code{classes}?}

\item{NAcolor}{Color used for \code{x} values not present in \code{classes}.}
}
\value{
A vector of colors of the same length as \code{x}.
}
\description{
Associates nominal values with a set of colors.
}
\examples{
classes=c("A","B","C")
colors=c("red","blue","green")
x= runif(1000) * 100
y= runif(length(x)) * 100
z= rep("A", length(x))
z[x*y > 50] = "B"
z[x*y > 500] = "C"
z[x*y > 5000] = "D"
plot(x, y, col=colorQual(z, classes, colors, NAcolor="grey"))
legend("topright", bg="white", pch=1, col=colors, legend=classes)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

