% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorQuant.r
\name{colorQuant}
\alias{colorQuant}
\title{Assign colors to numeric values}
\usage{
colorQuant(x, breaks, colors, lower = TRUE, warn = TRUE,
  NAcolor = "transparent")
}
\arguments{
\item{x}{Numeric vector.}

\item{breaks}{Break points for the colors. In the absence of more suitable
information, quantiles of \code{x} are often appropriate.}

\item{colors}{Vector of colors corresponding to \code{breaks}.}

\item{lower}{If \code{TRUE}, the data in \code{breaks} are interpreted as
lower boundaries of color intervals (thus the last element of \code{colors}
is ignored unless \code{max(x)} equals \code{max(breaks)}). If \code{TRUE},
the data in \code{breaks} are interpreted as upper interval boundaries
(thus the first element of \code{colors} is ignored unless \code{min(x)}
equals \code{min(breaks)}).}

\item{warn}{Warn if \code{x} contains values outside \code{range(breaks)}?}

\item{NAcolor}{Color assigned to \code{NA} values in \code{x}.}
}
\value{
A vector of colors of the same length as \code{x}.
}
\description{
Associates numeric values with a set of colors.
}
\examples{
x= runif(10)
x[3]= NA
x[5]= 2
breaks= c(0, .33, 0.66, 1)
colors= colorRampPalette(c("steelblue2","khaki2","brown"))(length(breaks))
low=FALSE
barplot(height=x, names.arg=1:length(x), ylim=c(0,1.2),
  col=colorQuant(x, breaks, colors, low))
abline(h=breaks, lty=3)
legend("top", bty="n", horiz=TRUE, fill=colorQuant(breaks, breaks, colors, low),
  legend=paste(ifelse(low, ">", "<"),breaks))
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

