% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeGrid.r
\name{makeGrid}
\alias{makeGrid}
\title{Depth grid for 1D models}
\usage{
makeGrid(dz0, dzMax, zMax, beta = 1)
}
\arguments{
\item{dz0}{Resolution at surface.}

\item{dzMax}{Resolution at suggested maximum depth.}

\item{zMax}{Suggested maximum depth.}

\item{beta}{Shape parameter. The default of 1 creates a grid whose resolution
  declines linearly with depth.}
}
\value{
A data frame with 1 row per layer and the following columns
\itemize{
  \item {dz} Thickness of layer.
  \item {zUp} Depth of layer's upper boundary.
   \item {zLw} Depth of layer's lower boundary.
}
}
\description{
Creates a grid with high resolution near surface and lower resolution at
greater depth.
}
\note{
The maximum depth of the generated grid (i.e. the lower boundary of the
 last layer) is guaranteed to be >= \code{zMax}.
 For a grid with high resolution near the surface but low resolution at
 greater depth one typically wants to set \code{beta} to a value > 1.
}
\examples{
# 0.3 m total depth, resolution of 2 mm near surface and 2 cm at depth
gr= makeGrid(dz0=0.002, dzMax=0.02, zMax=0.3, beta=2)
plot(gr$zUp, gr$dz, type="S", xlab="Depth", ylab="Thickness of layer")
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

