% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.r
\name{plot3d}
\alias{plot3d}
\title{Color Map of Dynamic Simulation Output}
\usage{
plot3d(obj, item, xlab = "time", ylab = "depth", rangeT = c(NA, NA),
  rangeX = c(NA, NA), partly = TRUE, timeconv = function(x) {    
  ISOdatetime(1970, 1, 1, 0, 0, 0) + x }, ...)
}
\arguments{
\item{obj}{Object created with \code{sim1D.create}.}

\item{item}{Name of a state variable or process (character string).}

\item{xlab}{Label for the time axis (x).}

\item{ylab}{Label for the depth axis (y).}

\item{rangeT}{Time range. Data outside this range are dropped.}

\item{rangeX}{Range of spatial coordinates (e.g. depths). Data outside this
range are dropped.}

\item{partly}{Logical. If \code{TRUE}, the result contains data for those
boxes/layers that are partly within the range \code{rangeX}. If
\code{FALSE}, only data for boxes/layers being entirely within
\code{rangeX} are returned.}

\item{timeconv}{A function to convert numeric times to class \code{POSIXct}.}

\item{...}{Further arguments to be passed to \code{fields::image.plot}. It is
often useful to pass \code{legend.mar} at least.}
}
\value{
\code{NULL}
}
\description{
Creates a color map from an object of class \code{sim1D}.
}
\note{
The function calls \code{fields::image.plot} which is possibly
  incompatible with the \code{layout} to arrange multiple plots.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

