% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotProfiles.r
\name{plotProfiles}
\alias{plotProfiles}
\title{Plot dynamic simulation output as vertical profiles}
\usage{
plotProfiles(name, dyn, gr, obs, xlab = "mmol/L", ylab = "depth",
  pos = "center", lty = rep(1, length(times)), col = 1:length(times), ...)
}
\arguments{
\item{name}{Name of the item (i.e. variable or process rate) to be plotted.}

\item{dyn}{Matrix generated by a call to \code{deSolve::ode}.}

\item{gr}{The grid used to create \code{dyn}. Must have been created by a
call to \code{\link{makeGrid}}.}

\item{obs}{Data frame of observations. See notes below.}

\item{xlab}{Label for the time axis (x).}

\item{ylab}{Label for the depth axis (y).}

\item{pos}{Keyword specifying the position of the legend.}

\item{lty}{Line types, 1 value or a vector as long as \code{times}.}

\item{col}{Line colors, 1 value or a vector as long as \code{times}.}

\item{...}{Additional arguments to be passed to \code{legend}.}
}
\value{
\code{NULL}
}
\description{
Plots the output of a dynamic 1D diagenetic model as vertical profiles.
}
\note{
The data frame of observations passed as \code{obs} needs to
  have 3 columns named 'time', 'depth', and 'value'. If the times of
  observation do not match exactly with the model output, data for the
  nearest time points available in \code{dyn} will be plotted.
  It is OK if the 'depth' and 'value' columns contain \code{NA} only and in
  that case, just the simulated data are plotted.
}
\examples{
gr= makeGrid(dz0=0.01, dzMax=0.02, zMax=0.2, beta=1)
times= 0:2
values= matrix(runif(length(times)*nrow(gr)), nrow=length(times))
dyn= cbind(times, values)
colnames(dyn)= c("time", paste("X",1:nrow(gr),sep="."))
obs= data.frame(time=c(1,1,2,2,2), depth=c(0.03, 0.04, 0.03, 0.04, 0.05))
obs$value= runif(nrow(obs))
plotProfiles("X", dyn, gr, obs)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

