% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeseries.r
\name{plotTimeseries}
\alias{plotTimeseries}
\title{Plot as time series}
\usage{
plotTimeseries(names, dyn, xlab = "", ylab = "", xlim = NULL,
  ylim = NULL, pos = "top", lty = 1:length(names), col = rep("black",
  length(names)), timeconv = function(x) {     ISOdatetime(1970, 1, 1, 0, 0,
  0) + x }, width = 4, height = 4, pointsize = 10, ofile = "out.svg",
  fmt = "svg", replace = TRUE, ...)
}
\arguments{
\item{names}{Names of items (i.e. variables or process rates) to plot.}

\item{dyn}{Matrix generated by a call to \code{deSolve::ode}.}

\item{xlab}{Label for the time axis (x).}

\item{ylab}{Label for the depth axis (y).}

\item{xlim}{Limits for the value axis (x). Computed if \code{NULL}.}

\item{ylim}{Limits for the depth axis (y). Computed if \code{NULL}.}

\item{pos}{Keyword specifying the position of the legend.}

\item{lty}{Line types, should be of the same length as \code{names}.}

\item{col}{Line colors, should be of the same length as \code{names}.}

\item{timeconv}{A function applied to the first column of \code{dyn} in order
to convert numeric times to values of class \code{POSIXct}.}

\item{width}{Plot width in inch.}

\item{height}{Plot height in inch.}

\item{pointsize}{Font size.}

\item{ofile}{Name of the outout file (with extension).}

\item{fmt}{Graphics format, either 'svg' or 'pdf'.}

\item{replace}{If \code{FALSE} existing output files result in an error.}

\item{...}{Additional arguments to be passed to \code{legend}.}
}
\value{
\code{NULL}
}
\description{
Plots the output of a dynamic model as a time series.
}
\examples{
times= 0:100
vars= c("a","b","c")
values= matrix(runif(length(times)*length(vars)), nrow=length(times))
dyn= cbind(times, values)
colnames(dyn)= c("time", vars)
plotTimeseries(c("a","c"), dyn, timeconv=function(x){x}, ofile=tempfile())
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

