% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seasonal.r
\name{seasonal}
\alias{seasonal}
\title{Uni-modal seasonality}
\usage{
seasonal(daynum, daymin, daymax, ymin, ymax)
}
\arguments{
\item{daynum}{Day of the year (value in range 1:366). Can be a vector.}

\item{daymin}{Day of the annual minimum, scalar.}

\item{daymax}{Day of the annual maximum, scalar.}

\item{ymin}{Value of the annual minimum, scalar.}

\item{ymax}{Value of the annual maximum, scalar.}
}
\value{
Numeric vector of the same length as \code{daynum}.
}
\description{
Approximation of seasonal variation by a (possibly asymmetric) sine function
}
\note{
A year is assumed to consist of 365 days. The result for day 366 is the
  same as for day 365. See the example for how this function can be used with
  values of class \code{POSIXct}.
}
\examples{
# Lake temperature: min. of 3 degree C at day 20, max. 23 degree C at day 210
times= seq(from=ISOdatetime(2000,1,1,0,0,0),
  to=ISOdatetime(2002,1,1,0,0,0), by=3600)
dayofyear= as.integer(format(times, "\%j"))
plot(times, seasonal(dayofyear, 20, 210, 3, 23), type="l", ylab="Temperature")
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

