% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1D.r
\name{sim1D.final}
\alias{sim1D.final}
\title{Final state of a 'sim1D' object}
\usage{
sim1D.final(obj)
}
\arguments{
\item{obj}{Object created with \code{sim1D.create}.}
}
\value{
A numeric matrix where rows represent boxes/layers. There is one
  column for each item and columns are named accordingly. Rows are unnamed
  but the following attributes are set:
\itemize{
  \item{\code{time} : } Scalar numeric representing the last time point for
    which data are stored in \code{obj}.
  \item{\code{coordinates} : } Spatial coordinates corresponding to the
    rows of the result. This is a matrix with two colums 'min' and 'max'.
}
}
\description{
Extract data from object of Class 'sim1D' for the very last time step
}
\examples{
gr <- makeGrid(dz0=0.01, dzMax=0.02, zMax=0.1, beta=1)
times <- 0:2
vars <- c("A","B")
pros <- c("X","Y")
values <- matrix(1:(length(times)*nrow(gr)*(length(vars)+length(pros))),
  nrow=length(times))
dyn <- cbind(times, values)
colnames(dyn) <- c("time", paste(rep(vars, each=nrow(gr)), 1:nrow(gr),sep="."),
  paste(rep(pros, each=nrow(gr)), 1:nrow(gr),sep="."))
obj <- sim1D.create(dyn, namesVars=vars, namesPros=pros, xMin=gr$zUp, xMax=gr$zLw)
sim1D.final(obj)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

