% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1D.r
\name{sim1D.query}
\alias{sim1D.query}
\title{Query an Object of Class 'sim1D'}
\usage{
sim1D.query(obj, item, rangeT = c(NA, NA), rangeX = c(NA, NA),
  partly = FALSE)
}
\arguments{
\item{obj}{Object created with \code{sim1D.create}.}

\item{item}{Name of a state variable or process (character string).}

\item{rangeT}{Time range. Data outside this range are dropped.}

\item{rangeX}{Range of spatial coordinates (e.g. depths). Data outside this
range are dropped.}

\item{partly}{Logical. If \code{TRUE}, the result contains data for those
boxes/layers that are partly within the range \code{rangeX}. If
\code{FALSE}, only data for boxes/layers being entirely within
\code{rangeX} are returned.}
}
\value{
A numeric matrix where rows represent time and columns represent
  boxes/layers. Row names are constructed from the time values. Column names
  are created by concatenating the boxes'/layers' minimum and maximum 
  coordinate, separated by comma.
}
\description{
Query an object of class 'sim1D' holding the output of a dynamic, 1D,
rodeo-based model.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

