% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyList.r
\name{simplifyList}
\alias{simplifyList}
\title{Transforms a list into a more compact format}
\usage{
simplifyList(L)
}
\arguments{
\item{L}{List to be transformed. All elements of \code{L} must share the same
type, length, dimension(s) and name(s). See the return value for supported
element types.}
}
\value{
The returned type depends on the type of the elements of \code{L}.
\itemize{
  \item{scalar case: } If each element of \code{L} is a scalar, the result
    is a vector of length \code{length(L)}.
  \item{vector case: } If each element of \code{L} is a vector, the result
    is a matrix. The number of rows equals \code{length(L)} and the number
    of columns is determined by the length of the vectors, i.e.
    \code{length(L[[1]])}.
  \item{matrix case: } If each element of \code{L} is a matrix, the result
    is an array with 3 dimensions. The length of dimension 1 equals the
    number of the matrices' rows, the length of dimension 2 equals the
    number of the matrices' columns, and the length of the 3rd dimension is
    \code{length(L)}.
}
}
\description{
Transforms a list of scalars, vectors, or matrices into a simplified
structure with one additional dimension (i.e. a vector, matrix, or 3D array).
}
\note{
An error is generated if the restrictions with respect to the elements
  of \code{L} are not met.
}
\examples{
# List of scalars
x= list(a=1, b=2, c=3)
print(simplifyList(x))  # same as unlist(x)
# List of vectors
x= list(a=1:2, b=2:3, c=3:4)
print(simplifyList(x))
# List of matrices
m= matrix(1:6, ncol=2)
colnames(m)= paste0("col",1:ncol(m))
rownames(m)= paste0("row",1:nrow(m))
x= list(a=m*1, b=m*2, c=m*3)
print(simplifyList(x))
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

