% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gof.r
\name{gof}
\alias{gof}
\title{Goodness-of-fit for a 1D dynamic, multi-component model}
\usage{
gof(dyn, gr, obs, f = function(sim, obs) {     mean((sim - obs)^2) })
}
\arguments{
\item{dyn}{Matrix generated by a call to \code{deSolve::ode}.}

\item{gr}{The grid used to create \code{dyn}. Must have been created by a
call to \code{\link{makeGrid}}.}

\item{obs}{Data frame of observations. See notes below.}

\item{f}{A function taking two numeric vectors with observed and simulated
  data as input. It should return a numeric vector (which can be of length 1).}
}
\value{
A list with as many elements as there are corresponding variables in
  \code{dyn} and \code{obs}. Each element contains the output of \code{f}.
  If, for a particular variable, no corresponding data are found (because
  of a missing overlap of depth and/or time interval), the list element for
  that variable is set to \code{NULL}.
  If there aren't any corresponding variables (due to a lack of matching
  names), the return value is a list of length 0.
}
\description{
Compares observation data with the model output.
}
\note{
The data frame of observations passed as \code{obs} needs to
  have 4 columns named 'time', 'depth', 'variable', and 'value'.
  Simulated values are assigned to the observation data based on a
  nearest-neighbor principle, i.e. data are picked from the nearest bin with
  respect to both time and depth.
}
\examples{
gr= makeGrid(dz0=0.01, dzMax=0.02, zMax=0.1, beta=1)
times= 0:2
vars= c("X","Y")
values= matrix(1:(length(times)*nrow(gr)*length(vars)), nrow=length(times))
dyn= cbind(times, values)
colnames(dyn)= c("time", paste("X",1:nrow(gr),sep="."),
  paste("Y",1:nrow(gr),sep="."))
obs= data.frame(time=c(1,1,2,2), depth=c(0.03, 0.04, 0.05, 0.04),
  variable=c("X","X","X","Y"), value=0)
fun= function(sim,obs) {c(bias=mean(sim-obs), rmse=mean((sim-obs)^2))}
gof(dyn, gr, obs, f=fun)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

