% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot3d.r
\name{plot3d}
\alias{plot3d}
\title{Color map of dynamic simulation output}
\usage{
plot3d(name, dyn, gr, xlab = "time", ylab = "depth",
  timeconv = function(x) {     ISOdatetime(1970, 1, 1, 0, 0, 0) + x }, ...)
}
\arguments{
\item{name}{Name of the item (i.e. variable or process rate) to be plotted.}

\item{dyn}{Matrix generated by a call to \code{deSolve::ode}.}

\item{gr}{The grid used to create \code{dyn}. Must have been created by a
call to \code{\link{makeGrid}}.}

\item{xlab}{Label for the time axis (x).}

\item{ylab}{Label for the depth axis (y).}

\item{timeconv}{A function applied to the first column of \code{dyn} in order
to convert numeric times to values of class \code{POSIXct}.}

\item{...}{Further arguments to be passed to \code{fields::image.plot}. It is
  often useful to pass \code{legend.mar} at least.}
}
\value{
\code{NULL}
}
\description{
Plots the output of a dynamic 1D diagenetic model.
}
\note{
The function calls \code{fields::image.plot} which is possibly
  incompatible with the \code{layout} to arrange multiple plots.
}
\examples{
gr= makeGrid(dz0=0.01, dzMax=0.02, zMax=0.2, beta=1)
times= 0:10
values= matrix(runif(length(times)*nrow(gr)), nrow=length(times))
dyn= cbind(times, values)
colnames(dyn)= c("time", paste("X",1:nrow(gr),sep="."))
plot3d("X", dyn, gr)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

