% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfiles.r
\name{plotProfiles}
\alias{plotProfiles}
\title{Plot dynamic simulation output as vertical profiles}
\usage{
plotProfiles(obj, item, obs, xlab = "", ylab = "depth", xlim = NULL,
  ylim = NULL, pos = "top", timeconv = function(x) {     ISOdatetime(1970,
  1, 1, 0, 0, 0) + x }, width = 4, height = 4, pointsize = 10,
  dir = ".", fmt = "svg", prefix = "profile", replace = TRUE, ...)
}
\arguments{
\item{obj}{Object created with \code{sim1D.create}.}

\item{item}{Name of a state variable or process (character string).}

\item{obs}{Data frame of observations. See notes below.}

\item{xlab}{Label for the time axis (x).}

\item{ylab}{Label for the depth axis (y).}

\item{xlim}{Limits for the value axis (x). Computed if \code{NULL}.}

\item{ylim}{Limits for the depth axis (y). Computed if \code{NULL}.}

\item{pos}{Keyword specifying the position of the legend.}

\item{timeconv}{A function to convert numeric times to class \code{POSIXct}.}

\item{width}{Plot width in inch.}

\item{height}{Plot height in inch.}

\item{pointsize}{Font size.}

\item{dir}{Output directory for generated graphics.}

\item{fmt}{Graphics format, either 'svg' or 'pdf'.}

\item{prefix}{Prefix put before the value of \code{name} when constructing
file names.}

\item{replace}{If \code{FALSE} existing output files result in an error.}

\item{...}{Additional arguments to be passed to \code{legend}.}
}
\value{
A vector holding the names of the generated graphics files.
}
\description{
Plots the output of a dynamic 1D diagenetic model as vertical profiles.
}
\note{
The data frame of observations passed as \code{obs} needs to
  have 3 columns named 'time', 'depth', and 'value'. If the times of
  observation do not match exactly with the model output, data for the
  nearest time points available in \code{dyn} will be plotted. Data in the
  'time' column must be of the same type as the return value of \code{timeconv}.
  It is OK if the 'depth' and 'value' columns contain \code{NA} only and in
  that case, just the simulated data are plotted.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

