% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.r
\name{seasonal}
\alias{seasonal}
\title{Uni-modal seasonality}
\usage{
seasonal(daynum, day1, day2, y1, y2)
}
\arguments{
\item{daynum}{Day of the year (value in range 1:366). Can be a vector.}

\item{day1}{First day with fixed value, scalar.}

\item{day2}{Second day with fixed value, scalar. Must be > \code{day1}.}

\item{y1}{Value at \code{day1}, scalar.}

\item{y2}{Value at \code{day2}, scalar.}
}
\value{
Numeric vector of the same length as \code{daynum}.
}
\description{
Approximation of seasonal variation by a (possibly asymmetric) sine function
}
\note{
A year is assumed to consist of 365 days. The result for day 366 is the
  same as for day 365. See the example for how this function can be used with
  values of class \code{POSIXct}.
}
\examples{
# Lake temperature: min. of 3 degree C at day 20, max. 23 degree C at day 210
times= seq(from=ISOdatetime(2000,1,1,0,0,0),
  to=ISOdatetime(2002,1,1,0,0,0), by=3600)
dayofyear= as.integer(format(times, "\%j"))
plot(times, seasonal(dayofyear, 20, 210, 3, 23), type="l", ylab="Temperature")
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

