% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1D.r
\name{sim1D.create}
\alias{sim1D.create}
\title{Create an Object of Class 'sim1D'}
\usage{
sim1D.create(dyn, model, xMin, xMax, keepItems = c(model$namesVars(),
  model$namesPros()), sep = ".", rangeT = c(NA, NA), rangeX = c(NA, NA),
  nsig = 2)
}
\arguments{
\item{dyn}{Numeric matrix output by the ODE solver
(e.g. \code{\link[deSolve]{ode}}).}

\item{model}{An object of the \code{\link[rodeo]{rodeo-class}} representing
the model used to create \code{dyn}.}

\item{xMin}{Minimum coordinates of the spatial boxes, e.g. top layer
boundaries in a vertical 1D model.}

\item{xMax}{Maximum coordinates of the spatial boxes, e.g. bottom layer
boundaries in a vertical 1D model.}

\item{keepItems}{Names of the items to be kept. Must be names of state
variables or processes existing in \code{model}.}

\item{sep}{The character used in the column names of \code{dyn} to separate
the variable/process name from the box/layer index.}

\item{rangeT}{Time range. Data outside this range are dropped.}

\item{rangeX}{Range of spatial coordinates (e.g. depths). Data outside this
range are dropped.}

\item{nsig}{Number of significant digits to retain (integer).}
}
\value{
An object of class 'sim1D'. Use \code{sim1D.query} to extract data
  from such objects.
}
\description{
Create an object of class 'sim1D' from the output of a dynamic, 1D,
rodeo-based model.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

